% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{allNAs}
\alias{allNAs}
\title{Check if vector is only NAs}
\usage{
allNAs(vec)
}
\arguments{
\item{vec}{Vector of values to be checked}
}
\value{
boolean TRUE or FALSE
}
\description{
Check if vector is only NAs
}
\examples{
allNAs(c(NA, NA, NA))
allNAs(c(NA,1,NA))
allNAs(c(1,2,3))
allNAs(c())
}
