% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{changeAnchorCoord}
\alias{changeAnchorCoord}
\title{Change the anchor coordinate}
\usage{
changeAnchorCoord(mat, anchorCoord = 0)
}
\arguments{
\item{mat}{A methylation matrix}

\item{amchorCoord}{The coordinate which will be set as the 0 position for relative coordinates 
(default=0)}
}
\value{
A methylation matrix in which the column names have been changed to indicate relative position
with reference to the anchor coordinate. e.g. a 500bp matrix centered around the TSS can have its column 
names changed from 0 to 500 range to -250 to 250 range by setting anchorCoord=250.
}
\description{
Change the anchor coordinate
}
