% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{combineCGGCgr}
\alias{combineCGGCgr}
\title{Convert Genomic ranges list to single genomic range with sample methylation and total counts in metadata}
\usage{
combineCGGCgr(methFreqGR, samples)
}
\arguments{
\item{methFreqGR}{A list (by sample) of a list (by C context) of genomic ranges for cytosine methylation frquency}

\item{samples}{a vector of sample names for which to extract the data}
}
\value{
A genomic ranges wtih all CG and GC positions found in the samples. The metadata columns contain methylation frequency (_M) and total read count (_T) for each sample
}
\description{
genomic ranges for CG and GC positions in each samples are combined and sorted. Then genomic ranges
from different samples are merged together putting methylation frequency (_M) and total read counts (_T) for each sample in the metadata.
}
