% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{convertGRtoRelCoord}
\alias{convertGRtoRelCoord}
\title{Convert Genomic Ranges to relative cooridnates}
\usage{
convertGRtoRelCoord(grs, winSize, anchorPoint = "middle")
}
\arguments{
\item{grs}{A GenomicRanges object to be converted to relative coordinates}

\item{winSize}{The size of the window you wish to create}

\item{anchorPoint}{One of "middle" or "start": the position from which numbering starts}
}
\value{
A GenomicRanges object with relative coordinate numbering
}
\description{
Convert a normal genomic ranges to one where the start and end are relative to some
anchor point - either the middle or the start of the genomic ranges (e.g. -250 to 250, or
0 to 500). The original start end and strand are stored in the metadata.
}
