% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{convertRelCoordtoGR}
\alias{convertRelCoordtoGR}
\title{Convert relative cooridnates to absolute genomic position}
\usage{
convertRelCoordtoGR(grsRelCoord, regionGR, anchorPoint = "middle")
}
\arguments{
\item{grsRelCoord}{A GenomicRanges object with one or more relative
coordinate ranges to be converted to absolute genomic positions.}

\item{regionGR}{A GenomicRanges object for the whole region to which the
grsRelCoord are relative to.}

\item{anchorPoint}{One of "middle" or "start": the position from which numbering starts}
}
\value{
A GenomicRanges object with absolute genomic position
}
\description{
Convert genomic ranges where the start and end are
 relative to some anchor point - either the middle or the start of
 the genomic ranges (e.g. -250 to 250, or 0 to 500) to a normal Genomic
 Ranges with absolute genomic positions.
}
