% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacBio_m6A.R
\name{fiberseqBedToBigwig}
\alias{fiberseqBedToBigwig}
\title{Convert fiberseq bed file to bigwig}
\usage{
fiberseqBedToBigwig(
  bedGR,
  genome = BSgenome.Celegans.UCSC.ce11::Celegans,
  ATpositionGR = NULL,
  minSubreadCov = 10,
  minReadCov = 5
)
}
\arguments{
\item{bedGR}{GRanges object of fiberseq output (bed) imported into R}

\item{genome}{BSgenome object for your organism with UCSC style seqinfo (Default
is C. elegans ce11)}

\item{ATpositionGR}{GRanges object with position of all As and Ts in genome}

\item{minSubreadCov}{Minimum coverage of the read with subreads (default=10)}

\item{minReadCov}{Minimum read coverage per genomic position (default=5)}
}
\value{
List of two GRanges: The first is the original bedGR but with additonal
metadata columns for fraction methylation per fiber. The second contains all the
A/T sites along the genome for which a valid fraction methylation could be calculated.
}
\description{
Processes data from bed file that is the output of
Andrew Stergachis fiberseq scripts where each row is a
fiber and the final columns indicate methylated positions
in that fiber. Output is: 1) the original GRanges with all invalide fibers removed
and with additional column denoting fraction methylation, 2) A new GRanges
with all valid A/T postions and their fraction methylation.
}
