% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{getAllSampleMetaMethFreq}
\alias{getAllSampleMetaMethFreq}
\title{get average methylation frequency from all matrices}
\usage{
getAllSampleMetaMethFreq(relCoordMats, samples, regionGRs, minReads = 10)
}
\arguments{
\item{relCoordMats}{A list (by sample) of lists (by regions) of methylation matrices which have been converted to relative coordinates}

\item{samples}{A list of samples to plot (same as sample names in allSampleMats)}

\item{regionGRs}{A genomic regions object with all regions for which matrices should be extracted (same as in allSampleMats). The metadata columns must contain a column called "ID" with a unique ID for each region.}

\item{minReads}{The minimal number of reads required in a matrix for average frequency to be calculated.}
}
\value{
Data frame with average methylation at relative coordinates extracted from all matrices for all samples
}
\description{
In order to do a metagene plot from matrices, the average methylation frequency from all
matrices with more reads than minReads is collected into a data frame
}
