% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{getFullMatrices}
\alias{getFullMatrices}
\title{Get full matrix of all positions in a window even if no C}
\usage{
getFullMatrices(matList, regionType, winSize = 500, workDir = ".")
}
\arguments{
\item{matList}{A table of paths to matrices which have the same columns}

\item{regionType}{A name for the type of region the matrices desribe}

\item{winSize}{The size (in bp) of the window containing the matrices}

\item{workDir}{The path to working directory}
}
\value{
A table of file paths to padded methylation matrices
}
\description{
In order to compare different promoters, we need to create a padded
matrix with NAs in positions in between Cs.
}
