% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{getMatrices}
\alias{getMatrices}
\title{Extract matrices into a simple list}
\usage{
getMatrices(methMats, regionName = c(), sampleName = c())
}
\arguments{
\item{methMats}{A table of filepaths to methylation matrices
(usually named by the genomic region they come from)}

\item{regionName}{A vector containing the names of regions of interest (default is all regions)}

\item{sampleName}{A vector containing the names of samples of interest (default is all samples)}
}
\value{
A simple, one level list of methylation matrices for the samples and regions of interest.
Matrices will be named with both the sample and the region name joined by "__"
}
\description{
Extract matrices into a simple list
}
\examples{
getMatrices(methMats)
#will return a simple list of all regions in all samples, changing the name of the matrices
getMatrices(methMats,regionName="WBGene00009234")
#will return a simple list of the matrices for that gene from all samples
getMatrices(methMats,regionName=c("WBGene00009234","WBGene00009621"))
#will return a simple list of the matrices for both those genes from all samples
getMatrices(methMats,sampleName="182_dSMFv002amp")
#will return a simple list of the matrices for all gene from this sample only
getMatrices(methMats,regionName="WBGene00009234",sampleName="182_dSMFv002amp")
#will return a single item list of a matrix for that gene in that sample.
}
