% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{getMatrixLog}
\alias{getMatrixLog}
\title{Read in an existing matrixLog, or create a new one}
\usage{
getMatrixLog(matrixLogFile, samples, regionGRs)
}
\arguments{
\item{matrixLogFile}{Name of matrix log file complete with relative or absolut path}

\item{samples}{Vector with names of samples to be processed}

\item{regionGRs}{A genomic regions object with all regions for which matrices should be extracted. The metadata columns must contain a column called "ID" with a unique ID for that region.}
}
\value{
A data frame with columns to record data about the single molecule matrices
}
\description{
Input requires matrixLog file name with correct path. If file exists it will be read
in, otherwise an empty data.frame will be created with the correct number of fields
for the samples (samples) and the regions (regionGRs). Teh function returns this
matrixLog data.frame
}
