% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{getMetaMethFreq}
\alias{getMetaMethFreq}
\title{Calculate methylation frequency at each position for metagene plots}
\usage{
getMetaMethFreq(matList, regionGRs, minReads = 50)
}
\arguments{
\item{matList}{A simple list of methylation matrices with names that match the regionsGRs object}

\item{regionGRs}{A genomicRanges object of the regions for which aggregate methylation frequency
will be calculated. the object must contain a metadata column called "ID" containing names that
match the methylation matrices in matList}

\item{minReads}{The minimal number of reads a matrix must have in order to be used (default=50)}
}
\value{
A long form  dataframe with four columns: "position" is the C position within the genomic Ranges,
"methFreq" is the frequency of methylation at that position, "ID" is the name of the region, "chr"
is the chromosome on which that region is present.
}
\description{
Calculate methylation frequency at each position for metagene plots
}
