% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{getMethFreqGR}
\alias{getMethFreqGR}
\title{Get methylation frequency genomic ranges}
\usage{
getMethFreqGR(baseFileName, pathToMethCalls, motifFile, minDepth = 5)
}
\arguments{
\item{baseFileName}{base file name used in MethylDackel extract command (assumes _CpG.bedgraph, _CHH.bedgraph and _CHG.bedgraph extensions were later added by MethylDackel)}

\item{pathToMethCalls}{Path to folder in which MethylDackel output is found}

\item{motifFile}{Path to GRanges object with all unique CG/GC/GCGorCGC motifs in genome}

\item{minDepth}{Minimum read depth. Positions with fewer than this number are discarded (default=5)}
}
\value{
Named list with CG, GC and C genomic ranges with metadata about methylation.
}
\description{
Takes MethylDackel methylation call bedgraph files for CpG CHH and CHG motifs and creates genomic ranges for CG, GC and all other Cs. Output is a list of these three genomic ranges objects.
}
