% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{getReadMatrix}
\alias{getReadMatrix}
\title{Extract methylation matrix from alignment file}
\usage{
getReadMatrix(bamFile, genomeFile, bedFile, region)
}
\arguments{
\item{bamFile}{String with path to bam file with alignments of reads to genome}

\item{genomeFile}{String with path to fasta file with genome sequence}

\item{bedFile}{String with path to .bed file with locations of Cs to be evaluated}

\item{region}{Genomic range of region for which to extract reads. It can also be a string
denoting the region, e.g. "X:8069396-8069886"}
}
\value{
A matrix with reads as row names and C positions as column names.
Every position in matrix has a value of 0 (methylated), 1 (not methylated) or
NA (undetermined)
}
\description{
Extract methylation matrix from alignment file
}
