% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{getRelativeCoord}
\alias{getRelativeCoord}
\title{Convert C position numbering from genomic to relative coordinates}
\usage{
getRelativeCoord(mat, regionGR, invert = F)
}
\arguments{
\item{mat}{A methylation matrix}

\item{regionGR}{A genomicRanges object of the region relative to which the new coordinates are caclulated}

\item{invert}{A logical variable to indicate if the region should be inverted (e.g. if it is on the negative strand). Default: FALSE}
}
\value{
A methylation matrix in which the column names have been changed from absolute genomic positions to relative
positions within the genomicRange regionGR
}
\description{
Convert C position numbering from genomic to relative coordinates
}
