% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{grlToDf}
\alias{grlToDf}
\title{Convert Genomic ranges list to long data frame for plotting}
\usage{
grlToDf(methFreqGR, samples, Ctype)
}
\arguments{
\item{methFreqGR}{A list (by sample) of a list (by C context) of genomic ranges for cytosine methylation frquency}

\item{samples}{a vector of sample names for which to extract the data}

\item{Ctype}{The type of sequence context for the cytosine _("CG","GC" or "C")}
}
\value{
A long-form data frame with methylation frequency and counts at different sites in different samples
}
\description{
Methylation frequency data stored in  a list(by sample) of list (by C context) of genomic ranges is extracted into a data frame. By specifying a context type ("CG","GC" or "C") this type of data is extracted for all samples listed in "samples" can converted to a dataframe.
}
