% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacBio_m6A.R
\name{makeATgrObj}
\alias{makeATgrObj}
\title{Make GRanges object of all As and Ts in genome}
\usage{
makeATgrObj(genome = BSgenome.Celegans.UCSC.ce11::Celegans)
}
\arguments{
\item{genome}{BSgenome object for your organism (default is C.elegans)}
}
\value{
GRanges object
}
\description{
Make GRanges object of all As and Ts, with runs of A/Ts
unmerged.
}
