% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{makeCGorGCbed}
\alias{makeCGorGCbed}
\title{Create bed file of CpG or GpC motifs}
\usage{
makeCGorGCbed(genomeFile, CGorGC)
}
\arguments{
\item{genomeFile}{String with path to fasta file with genome sequence}

\item{CGorGC}{String with motif to be found (either "CG" or "GC")}
}
\value{
A bed file with locations of the C in all occurrences of this motif on both strands will be written to
the same directory as genomeFile with the extension .CpG.bed or .GpC.bed
}
\description{
Create bed file of CpG or GpC motifs
}
