% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{makeDirs}
\alias{makeDirs}
\title{Make directories}
\usage{
makeDirs(path, dirNameList = c())
}
\arguments{
\item{path}{String with path to where the directories should be made}

\item{dirNameList}{Vector of strings with names of directories to create (can include multilevel directories)}
}
\value{
Creates the directories listed in dirNameList
}
\description{
Make directories
}
\examples{
makeDirs(path=".",dirNameList=c("/txt","/rds/sample1"))
}
