% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{matToGR}
\alias{matToGR}
\title{Convert methylation matrix to genomic ranges}
\usage{
matToGR(mat, gr)
}
\arguments{
\item{mat}{Methylation matrix with reads x Cposition}

\item{gr}{GRanges object for which the methylation matrix was made}
}
\value{
genomic ranges of Cpositions  with mcols containing the reads as columns
}
\description{
Convert methylation matrix to genomic ranges
}
