% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{mergeSampleMats}
\alias{mergeSampleMats}
\title{Merge tables listing methylation matrices for different samples}
\usage{
mergeSampleMats(path, regionType, samples, deleteSplitFiles = F)
}
\arguments{
\item{path}{Path to working directory in which csv subdirectory exists.}

\item{regionType}{A collective name for this list of regions (e.g TSS or amplicons). It is used in naming the files}

\item{samples}{A list of sample names to merge (used in the name of the file)}

\item{deleteSplitFiles}{Logical value to determine if individual sample files
should be deleted (defualt=F).}
}
\value{
Table of merged samples
}
\description{
Uses file naming convention of MatrixLog_regionType_sampleName.csv to
merge tables of  methylation matrices that were processed separately for
each sample. The function will merge all the samples listed in samples and
then will also delete the original split files
}
