% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatsData.R
\docType{data}
\name{methMats}
\alias{methMats}
\title{Methylation matrix from dSMF analysis}
\format{
An object of class list (it is a nested list)
}
\usage{
data(methMats)
}
\description{
Data from dSMF experiment from two strains of C. elegans: wildtype N2 Bristol, and fog-2 mutant
(labelled 182). There are two samples for each strain.
The data is a list of lists: a list of samples, and each sample contains a list of methylation
matrices centered +-250bp around transcriptional start sites (TSSs) of C. elegans genes. The names of
the matrices in the list are the Wormbase gene names. The column names in the matrices are the genomic
cooridnates of cytosines that form part of a CpG or GpC motif. The rows of the matrices are pairs of
sequence reads generated by sequencing 400-500bp amplicons with 2x250 paired end Illumina sequencing.
Rownames are the read names.
}
\examples{
data(methMats)
}
\references{
Semple.J.I. & Meister P.
}
\keyword{datasets}
