% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{parseMatchString}
\alias{parseMatchString}
\title{Clean up match string from pileup file}
\usage{
parseMatchString(match)
}
\arguments{
\item{match}{Match string from pileup file.}
}
\value{
A cleaned up match string with no indels or start/end line info
}
\description{
Clean up match string from pileup file
}
