% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{pileupToConversionStatus}
\alias{pileupToConversionStatus}
\title{Extract methylation calls from single line of pileup file}
\usage{
pileupToConversionStatus(pileupLine)
}
\arguments{
\item{pileupLine}{One line data frame from pileup file. Columns have been named: c("chr","start","ref","count","matches","BQ","reads")}
}
\value{
A data frame with C position, read name and methylation call  0 (not methylated), 1 (methylated)
}
\description{
Extract methylation calls from single line of pileup file
}
