% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{plotAllMatrices}
\alias{plotAllMatrices}
\title{Plot a list of list of single molecule matrices}
\usage{
plotAllMatrices(allSampleMats, samples, regionGRs, featureGRs, regionType,
  featureLabel = "TSS", maxNAfraction = 0.2, withAvr = FALSE,
  includeInFileName = "", drawArrow = TRUE, workDir = ".")
}
\arguments{
\item{allSampleMats}{A list (by sample) of lists (by regions) of methylation matrices}

\item{samples}{A list of samples to plot (same as sample names in allSampleMats)}

\item{regionGRs}{A genomic regions object with all regions for which matrices should be extracted (same as in allSampleMats). The metadata columns must contain a column called "ID" with a unique ID for each region.}

\item{featureGRs}{A genomic regions object for features (such as TSS) to be plotted. Feature must be identified with the same ID as the regionGRs}

\item{regionType}{A collective name for this list of regions (e.g TSS or amplicons). It will be used in naming the output directories}

\item{featureLabel}{A string with a label for the feature to be added to the plot (default="TSS")}

\item{maxNAfraction}{Maximual fraction of CpG/GpC positions that can be undefined (default=0.2)}

\item{withAvr}{Boolean value: should single molecule plots be plotted together with the average profile (default=FALSE)}

\item{includeInFileName}{String to be included at the end of the plot file name, e.g. experiment name (default="")}

\item{drawArrow}{Boolean: should the feature be drawn as an arrow or just a line? (default=TRUE)}

\item{workDir}{Path to working directory}
}
\value{
Plots are written to plots directory
}
\description{
This function takes a list (by sample) of a list (by genomic region) of
methylation matrices and produces single molecule plots for each amplicon with
four samples per page.
}
