% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{plotDSMFmetageneDF}
\alias{plotDSMFmetageneDF}
\title{Plot metagene by sample}
\usage{
plotDSMFmetageneDF(metageneDF, maxPoints = 10000)
}
\arguments{
\item{metageneDF}{ata frame with average methylation at relative coordinates extracted from all matrices for all samples}

\item{maxPoints}{The maximum number of points to plot per sample. To avoid large files with too much overplotting, the defualt limit is set to 10000. Larger dataframes will be randomly sub-sampled}
}
\value{
A ggplot2 plot object
}
\description{
Plots metagene methylation frequency from dataframe extracted from matrices.
}
