% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{plotSingleMolecules}
\alias{plotSingleMolecules}
\title{Single molecule plot of a methylation matrix}
\usage{
plotSingleMolecules(
  mat,
  regionName,
  regionGRs,
  featureGRs = NULL,
  myXlab = "CpG/GpC position",
  featureLabel = "TSS",
  drawArrow = TRUE,
  title = NULL,
  baseFontSize = 12,
  maxNAfraction = 0.2,
  segmentSize = 3,
  colourChoice = list(low = "blue", mid = "white", high = "red", bg = "white", lines =
    "black"),
  colourScaleMidpoint = 0.5,
  doClustering = T
)
}
\arguments{
\item{mat}{A methylation matrix}

\item{regionName}{The name of the region the matrix is taken from should match on of the IDs in regionGRs}

\item{regionGRs}{A genomicRanges object which includes the region which the mat matrix provides the data for.
The object must contain a metadata column called "ID" containing names that match the methylation
matrices in mat}

\item{featureGRs}{A genomicRanges object denoting features to be plotted such as the TSS}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS
(default="TSS)}

\item{drawArrow}{Boolean: should the feature be drawn as an arrow or just a line? (default=TRUE)}

\item{title}{A title for the plot (default will be the name of the region, the chr and strand on which
the region is present)}

\item{baseFontSize}{The base font for the plotting theme (default=12 works well for 4x plots per A4 page)}

\item{maxNAfraction}{Maximual fraction of CpG/GpC positions that can be undefined (default=0.2)}

\item{segmentSize}{Length of colour segment denoting methylation site}

\item{colourChoice}{A list of colours for colour pallette. Must include
values for "low", "mid", "high" and "bg" (background) and "lines".}
}
\value{
A ggplot2 plot object
}
\description{
Single molecule plot of a methylation matrix
}
