% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{plotSingleMoleculesWithAvr}
\alias{plotSingleMoleculesWithAvr}
\title{Single molecule plot of a methylation matrix with average methylation frequency}
\usage{
plotSingleMoleculesWithAvr(mat, regionName, regionGRs, featureGRs,
  myXlab = "CpG/GpC position", featureLabel = "TSS", title = NULL,
  baseFontSize = 11)
}
\arguments{
\item{mat}{A methylation matrix}

\item{regionName}{The name of the region the matrix is taken from should match on of the IDs in regionGRs}

\item{regionGRs}{A genomicRanges object which includes the region which the mat matrix provides the data for.
The object must contain a metadata column called "ID" containing names that match the methylation
matrices in mat}

\item{featureGRs}{A genomicRanges object denoting features to be plotted such as the TSS}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS
(default="TSS)}

\item{title}{A title for the plot (default will be the name of the region, the chr and strand on which
the region is present)}

\item{baseFontSize}{The base font for the plotting theme (default=11 works well for 4x plots per A4 page)}
}
\value{
A ggplot2 plot object
}
\description{
Single molecule plot of a methylation matrix with average methylation frequency
}
