% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{poorBisulfiteConversion}
\alias{poorBisulfiteConversion}
\title{Find bisulfite conversion rate of Cs in non-methylated context}
\usage{
poorBisulfiteConversion(bamFile, genomeFile, bedFileC, bedFileG, regionGR)
}
\arguments{
\item{bamFile}{String with path to bam file with alignments of reads to genome}

\item{genomeFile}{String with path to fasta file with genome sequence}

\item{bedFileC}{String with path to .bed file with locations of Cs to be evaluated (for forward strand calls)}

\item{bedFileG}{String with path to .bed file with locations of Gs to be evaluated (for reverse strand calls)}

\item{regionGR}{Genomic range of region for which to extract reads. It can also be a string
denoting the region, e.g. "X:8069396-8069886"}
}
\value{
A data frame with the names of the reads, the count of the number of informative Cs
per region (not NAs), the maximum number of possible Cs in the regon, and the fraction of
the informative Cs which have been bisulfite converted
}
\description{
Use bedfile with positons of Cs that are in non-methylated context to obtain stats
about the number of informative Cs and conversion status of Cs per read
}
