% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{rbindMatrixList}
\alias{rbindMatrixList}
\title{#' Extract matrices into a simple list
#'
#' @param methMats A table of filepaths to methylation matrices
#' (usually named by the genomic region they come from)
#' @param regionName A vector containing the names of regions of interest (default is all regions)
#' @param sampleName A vector containing the names of samples of interest (default is all samples)
#' @return A simple, one level list of methylation matrices for the samples and regions of interest.
#' Matrices will be named with both the sample and the region name joined by "__"
#' @examples
#' getMatrices(methMats)
#' #will return a simple list of all regions in all samples, changing the name of the matrices
#' getMatrices(methMats,regionName="WBGene00009234")
#' #will return a simple list of the matrices for that gene from all samples
#' getMatrices(methMats,regionName=c("WBGene00009234","WBGene00009621"))
#' #will return a simple list of the matrices for both those genes from all samples
#' getMatrices(methMats,sampleName="182_dSMFv002amp")
#' #will return a simple list of the matrices for all gene from this sample only
#' getMatrices(methMats,regionName="WBGene00009234",sampleName="182_dSMFv002amp")
#' #will return a single item list of a matrix for that gene in that sample.
#' @export
getMatrices<-function(methMats,regionName=c(),sampleName=c()) {
  idx<-NULL
  # extracts a simple list of matrices. It is possible to also subset by regionName or sampleName
  # each matrix will have a name composed of sampleName__regionName
  newMats<-list()
  if(length(sampleName)==0) {
    sampleName<-unique(methMats$sample)
  }
  if(length(regionName)==0) {
    regionName<-unique(methMats$region)
  }
  s<-methMats$sample %in% sampleName
  r<-methMats$region %in% regionName
  idx<-r*s
  for (i in which(idx==1)) {
    mat<-readRDS(methMats$filename[i])
    newName<-paste0(methMats[idx,"sample"][i], "__", methMats[idx,"region"][i])
    newMats[[newName]]<-mat
  }
  return(newMats)
}
Merge multiple matrices into a single matrix}
\usage{
rbindMatrixList(matList)
}
\arguments{
\item{matList}{A table of paths to matrices which have the same columns}
}
\value{
A single methylation matrix
}
\description{
This function takes a list of matrices with the same number of columns and merges
}
