% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doparImport.R, R/magitterImport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%dopar\%}
\alias{\%>\%}
\title{import dopar function}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{foreach}{\code{\link[foreach]{\%dopar\%}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

