% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSgrData.R
\docType{data}
\name{TSS}
\alias{TSS}
\title{Genomic Ranges for methMats data TSS}
\format{
An object of class genomicRanges
}
\usage{
data(TSS)
}
\description{
Genomic ranges object for the transcription start site (TSS) that is within
the regions for which there is methylation matrix data in methMats data object
The metadata column "ID" contains the gene name for each TSS. This genomic ranges
object is of width 1 (just the TSS), and it should be extended to +- 250bp to define
the full region for which the methMats have been defined.
}
\examples{
data(TSS)
}
\references{
Semple.J.I. & Meister P.
}
\keyword{datasets}
