% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacBio_m6A.R
\name{fiberseqBedToBigwig}
\alias{fiberseqBedToBigwig}
\title{Convert fiberseq bed file to bigwig}
\usage{
fiberseqBedToBigwig(
  bedGR,
  genome = BSgenome.Celegans.UCSC.ce11::Celegans,
  ATpositionGR = NULL,
  minSubreadCov = 10,
  minReadCov = 5
)
}
\arguments{
\item{bedGR}{GRanges object of fiberseq output (bed) imported into R}

\item{genome}{BSgenome object for your organism with UCSC style seqinfo}

\item{ATpositionGR}{GRanges object with position of all As and Ts in genome}

\item{minSubreadCov}{Minimum coverage of the read with subreads (default=10)}

\item{minReadCov}{Minimum read coverage per genomic position (default=5)}
}
\value{
GRanges with score column containing % methylation
along genome
}
\description{
Processes data from bed file that is the output of
Andrew Stergachis fiberseq scripts where each row is a
fiber and the final columns indicate methylated positions
in that fiber
}
