% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2methMats.R
\name{getSingleMoleculeMatrices}
\alias{getSingleMoleculeMatrices}
\title{Extract list of methylation matrices}
\usage{
getSingleMoleculeMatrices(
  sampleTable,
  genomeFile,
  regionGRs,
  regionType,
  genomeMotifGR,
  minConversionRate = 0.8,
  maxNAfraction = 0.2,
  bedFilePrefix = NULL,
  path = ".",
  convRatePlots = FALSE,
  nThreads = 1,
  samtoolsPath = ""
)
}
\arguments{
\item{sampleTable}{Table with FileName column listing the full path to bam files belonging to the samples listed in the SampleName column}

\item{genomeFile}{String with path to fasta file with genome sequence}

\item{regionGRs}{A genomic regions object with all regions for which matrices should be extracted. The metadata columns must contain a column called "ID" with a unique ID for that region.}

\item{regionType}{A collective name for this list of regions (e.g TSS or amplicons)}

\item{genomeMotifGR}{A GenomicRanges object with a unique set of non-overlapping CG, GC and GCGorCGC sites}

\item{minConversionRate}{Minimal fraction of Cs from a non-methylation context that must be converted to Ts for the read to be included in the final matrix (default=0.8)}

\item{maxNAfraction}{Maximual fraction of CpG/GpC positions that can be undefined (default=0.2)}

\item{bedFilePrefix}{The full path and prefix of the bed file for C, G, CG and GC positions in the genome (i.e path and name of the file without the ".C.bed",".G.bed", ".CG.bed" or ".GC.bed" suffix). Defulat is NULL and assumes the bed file are in the same location as the genome sequence file.}

\item{path}{Path for output. "plots", "csv" and "rds" directories will be created here. Default is current directory.}

\item{convRatePlots}{Boolean value: should bisulfite conversion rate plots be created for each region? (default=FALSE)}

\item{nThreads}{number of threads for parallelisation}

\item{samtoolsPath}{Path to samtools executable (default="") if not in unix $PATH}
}
\value{
A list (by sample) of lists (by regions) of methylation matrices
}
\description{
Input requires a sampleTable with two columns: FileName contains the path to a bam file of
aligned sequences. SampleName contains the name of the sample. The function will return a
list of all samples and for each sample a list of methylation matrices for all regions listed
in regionGRs. The data structure is as follows:
[1] sample1
    [1] region1: matrix of reads x Cpositions
    [2] region2: matrix of reads x Cpositions
[2] sample2
    [1] region1: matrix of reads x Cpositions
    [2] region2: matrix of reads x Cpositions
Matrices contain values between 0 (not methylated) and 1 (methylated), or NA (undefined)
}
