% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methMatrixManipulation.R
\name{rbindMatrixList}
\alias{rbindMatrixList}
\title{Merge multiple matrices into a single matrix}
\usage{
rbindMatrixList(matList)
}
\arguments{
\item{matList}{A table of paths to matrices which have the same columns}
}
\value{
A single methylation matrix
}
\description{
This function takes a list of matrices with the same number of columns and merges
}
